requirements.txt
aiogram==3.15.0
asyncio==3.4.3
openai==1.59.3
python-dotenv==1.0.1
aiosqlite==0.20.0
loguru==0.7.3